+++ conanfile.py
@@ -23,8 +23,8 @@ class JSONSettingsConan(ConanFile):
         self.options["CaeserCypherEncryptionAdapter"].gtest = self.options.gtest
 
     def requirements(self):
-        self.requires("EncryptionAdapterInterface/1.0.5@systelab/stable")
-        self.requires("CaeserCypherEncryptionAdapter/1.0.4@systelab/stable")
+        self.requires("EncryptionAdapterInterface/1.1.0@systelab/stable")
+        self.requires("CaeserCypherEncryptionAdapter/1.1.0@systelab/stable")
         self.requires("RapidJSONAdapter/1.1.6@systelab/stable")
         if self.options.boost == "1.66.0":
             self.requires("boost/1.66.0@conan/stable")
@@ -34,7 +34,7 @@ class JSONSettingsConan(ConanFile):
             self.requires(("boost/%s") % self.options.boost)
 
     def build_requirements(self):
-        self.build_requires("EncryptionAdapterTestUtilities/1.0.5@systelab/stable#12f1bbb3ba384dc2e6c06c9d4072e7ff")
+        self.build_requires("EncryptionAdapterTestUtilities/1.1.0@systelab/stable")
         self.build_requires("JSONAdapterTestUtilities/1.1.5@systelab/stable")
         if self.options.gtest == "1.7.0":
             self.build_requires("gtest/1.7.0@systelab/stable")